[FEATURE] Improved link browser styles
The record list in the link browser is simplified with this patch. All
obsolete controls like the search box or the field selector are
disabled.

Additionally the fixed width of the title column is disabled by
registering the linkhandler Extension as a skin and overwriting the
CSS styles.

Finally the plus icon was removed from the record list since it does
not make sense here.