[FEATURE] Fallbacks for determining the current PID
The getCurrentPageUid() method in the Configuration manager is
enhanced and will now use either the first site root or the
first template root page that is found as fallback if other
methods fail to retrieve the current page ID.

This is a workaround for the invalid link handling in the tab
switch of the element browser in TYPO3.

Thanks to tritumRz for the initial pull request!

Resolves: #3