[CLEANUP] Remove old hook class
The class that used to hook into the RecordList display to filter
out translations cannot be used anymore as we don't rely on our
own RecordList class anymore. Thus it is not possible to know
in which context the list view is being displayed and act
only when in the LinkBrowser.

Maybe some solution can be found in the future, but for now
this feature is removed.